feat: complete rust ui state machine
